OpenAI_hello_agent/asynch/hello_agent/pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.11",
    "python-dotenv>=1.1.0",
]
